* 7.4:
  [Messenger] fix tests using Redis
  fix intl tests PHPUnit command
  chore: PHP CS Fixer - restore PHP / PHPUnit rulesets

+++ Tests/Transport/DoctrinePostgreSqlIntegrationTest.php
@@ -47,6 +47,9 @@ class DoctrinePostgreSqlIntegrationTest extends TestCase
 
     protected function tearDown(): void
     {
+        if (!isset($this->driverConnection)) {
+            return;
+        }
         $this->driverConnection->createSchemaManager()->dropTable('queue_table');
         $this->driverConnection->close();
     }
