+++ Tests/Transport/DoctrinePostgreSqlRegularIntegrationTest.php
@@ -87,6 +87,9 @@ class DoctrinePostgreSqlRegularIntegrationTest extends TestCase
 
     protected function tearDown(): void
     {
+        if (!isset($this->driverConnection)) {
+            return;
+        }
         $this->driverConnection->createSchemaManager()->dropTable('queue_table');
         $this->driverConnection->close();
     }
